Add gage and level system

+++ Assets/Scripts/Character/CharacterVisualManage.cs
@@ -8,11 +8,12 @@ public class CharacterVisualManage : MonoBehaviour
     [SerializeField] private Animator _faceAnim = null;
     int _thisLev = 0;
 
-    public void SetLevel(uint level)
+    public void SetLevel(int level)
     {
-        this._thisLev = (int)level;
+        this._thisLev = level;
         _anim.SetFloat("level", level * 0.25f);
-        _faceAnim.SetLayerWeight((int)level, 1);
+        _faceAnim.SetLayerWeight(level, 1);
+
         _faceAnim.SetFloat("temper", 0);
         _faceAnim.SetFloat("water", 0);
     }
